+++ src/Attractors.jl
@@ -8,6 +8,8 @@ module Attractors
 end Attractors
 
 using Reexport
+
+
 @reexport using StateSpaceSets
 @reexport using DynamicalSystemsBase
 
@@ -18,8 +20,11 @@ include("basins/basins.jl")
 include("continuation/basins_fractions_continuation_api.jl")
 include("boundaries/edgetracking.jl")
 include("deprecated.jl")
+include("tipping/tipping.jl")
 
 # Visualization (export names extended in the extension package)
 include("plotting.jl")
+
+# minimal fatal shock algo
 
 end # module Attractors