+++ src/basins/basins.jl
@@ -1,3 +1,2 @@
 include("basins_utilities.jl")
 include("fractality_of_basins.jl")
-include("tipping_probabilities.jl")